added settings design and remember me (login)
added logout functionality,

+++ src/app/pages/login/login.page.ts
@@ -3,6 +3,7 @@ import { Router } from '@angular/router';
 import { LoginData } from 'src/app/interfaces/login-data';
 import { ApiService } from 'src/app/services/api.service';
 import { Storage } from '@capacitor/storage';
+import { LoadingController } from '@ionic/angular';
 
 @Component({
   selector: 'app-login',
@@ -14,15 +15,52 @@ export class LoginPage implements OnInit {
   email: string;
   password: string;
 
-  constructor(private service: ApiService, private router: Router) {}
+  constructor(private service: ApiService, private router: Router, private loadingCtrl: LoadingController) {}
 
   ngOnInit() {
+    this.rememberLogin();
+  }
+
+  async rememberLogin() {
+    const refreshToken = await Storage.get({ key: 'refresh_token' });
+
+    if(refreshToken.value != null) {
+      // eventually solve with loadingScreen (animated)
+      const loading = await this.loadingCtrl.create({
+        translucent: true
+      });
+
+      loading.present();
+
+      this.service.rememberLogin(refreshToken.value)
+        .subscribe(
+          async (data) => {
+
+            // remove old tokens
+            await Storage.remove({ key: 'access_token' });
+            await Storage.remove({ key: 'refresh_token' });
+
+            await Storage.set({key: 'access_token', value: 'Bearer ' + data.data.tokens.access_token});
+            await Storage.set({key: 'refresh_token', value: data.data.tokens.refresh_token});
+
+            loading.dismiss();
+
+            this.router.navigate(['/tabs/tab1']);
+          },
+          async (error) => {
+            // delete possible old unvalid tokens
+            await Storage.remove({ key: 'access_token' });
+            await Storage.remove({ key: 'refresh_token' });
+
+            loading.dismiss();
+            console.log(error);
+          }
+        );
+    }
   }
 
   async login() {
 
-    await Storage.remove({ key: 'access_token' });
-    await Storage.remove({ key: 'refresh_token' });
     /*
     await Storage.remove({ key: 'name' });
 
@@ -34,22 +72,36 @@ export class LoginPage implements OnInit {
     console.log(await Storage.get({ key: 'name' }));
     */
 
-    const loginData: LoginData = {
-      email: this.email,
-      password: this.password
-    };
+    if(this.email.match(
+      // eslint-disable-next-line max-len
+      /^(([^<>()[\]\\.,;:\s@\"]+(\.[^<>()[\]\\.,;:\s@\"]+)*)|(\".+\"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/
+    )) {
 
-    this.service.login(loginData).subscribe(async (data) => {
+      // unecessary?
+      await Storage.remove({ key: 'access_token' });
+      await Storage.remove({ key: 'refresh_token' });
 
-      // store tokens in local database
-      await Storage.set({key: 'access_token', value: data.tokens.access_token});
-      await Storage.set({key: 'refresh_token', value: data.tokens.refresh_token});
+      const loginData: LoginData = {
+        email: this.email,
+        password: this.password
+      };
 
-      this.router.navigate(['/tabs/tab1']);
-    },
-    (error) => {
-      console.log(error);
-    });
+      this.service.login(loginData).subscribe(async (data) => {
+
+        // store tokens in local database
+        await Storage.set({key: 'access_token', value: 'Bearer ' + data.data.tokens.access_token});
+        await Storage.set({key: 'refresh_token', value: data.data.tokens.refresh_token});
+
+        this.router.navigate(['/tabs/tab1']);
+      },
+      (error) => {
+        console.log(error.error.message);
+      });
+
+    }
+    else {
+      console.log('Invalid input');
+      // TODO implement alert
+    }
   }
-
 }
